+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * Allow restriction to only one package manager for each sub-command.
   Closes #12.
 * Differentiate packages names and IDs. Closes #11.
+* Sort list of outdated packages by lower-cased package names first.
 * Fix parsing of multiple versions of ``cask`` installed packages.
 * Fix lexicographical sorting of ``brew`` and ``cask`` package versions.
 * Fix fall-back to iterative full upgrade command.